gallium/radeon: Make va_offset 64 bits wide.
Otherwise we'd wrap around after 32 bits. The kernel currently limits GPU
virtual address space to 4GB anyway, but that will probably change sooner or
later, and this would result in confusing error messages when running out of
virtual address space even now.

+++ src/gallium/winsys/radeon/drm/radeon_drm_bo.c
@@ -120,7 +120,7 @@ struct radeon_bomgr {
 
     /* is virtual address supported */
     bool va;
-    unsigned va_offset;
+    uint64_t va_offset;
     struct list_head va_holes;
 };
